+++ image_search/search_worker.h
@@ -53,9 +53,9 @@ class SearchWorker{
     //Find approximate KNN, this is supposed to be much faster than exact KNN when k is large.
     size_t search_K_approximate_nearest_neighbors(BinaryCode &code);
     void search_R_neighbors(int r, uint32_t search_index, 
-        std::vector<int> &knn_candidates);
+        std::vector<uint64_t> &knn_candidates);
     void enumerate_entry(uint32_t curr, int len, int rr, HashIndex &idx, 
-        std::vector<int> &knn_candidates);
+        std::vector<uint64_t> &knn_candidates);
     
     //try to map the memory space of bitmap deamon to local memory.
     bool connect_bitmap_deamon(unsigned long long size = ((unsigned long long )1 << 32) / 8 * 4);
